# crates/dark-wallet-rest/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "dark-wallet-rest",
    "description": "REST wallet daemon for dark, an Ark protocol server for Bitcoin L2.\n\nExposes a curated subset of the gRPC surface as HTTP + JSON, with SSE for server-streamed events. Signer-path RPCs (tree nonces, tree signatures, forfeit signing) remain gRPC-only by design.\n\nAll endpoints return JSON. Binary fields (PSBTs, pubkeys, signatures) are hex-encoded strings unless noted. Amounts are denominated in satoshis.",
    "contact": {
      "name": "Lobby",
      "email": "lobby.clawy@gmail.com"
    },
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/intents": {
      "post": {
        "tags": [
          "intents"
        ],
        "summary": "Register a round-settlement intent",
        "description": "Registers a BIP-322-signed intent for the next round. The body carries the proof PSBT (hex) and a canonical JSON `message`; optionally a `delegate_pubkey` (hex compressed) for delegated submission.",
        "operationId": "register_intent",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterIntentRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RegisterIntentResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Malformed request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/intents/{id}": {
      "delete": {
        "tags": [
          "intents"
        ],
        "summary": "Delete a registered intent",
        "operationId": "delete_intent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Intent id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Deleted"
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/intents/{id}/confirm": {
      "post": {
        "tags": [
          "intents"
        ],
        "summary": "Confirm participation in the current batch",
        "description": "Called after a `BatchStarted` event to acknowledge the VTXO tree.",
        "operationId": "confirm_intent",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Intent id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConfirmRegistrationRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Confirmed"
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/intents/{id}/fee": {
      "post": {
        "tags": [
          "intents"
        ],
        "summary": "Estimate the fee for a proposed intent",
        "description": "Takes the input VTXO ids and output destinations and returns the server's fee estimate. The path `{id}` is unused today but kept for symmetry with per-intent endpoints.",
        "operationId": "estimate_fee",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Intent id — unused",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EstimateIntentFeeRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Fee estimate",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EstimateIntentFeeResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/txs": {
      "post": {
        "tags": [
          "txs"
        ],
        "summary": "Submit a signed Ark virtual tx",
        "description": "Submits a hex-encoded signed Ark tx to the server. The server returns the assigned Ark txid. Call `/v1/txs/{id}/finalize` with the final checkpoint transactions to complete the flow.",
        "operationId": "submit_tx",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/SubmitTxRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Accepted",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SubmitTxResponseDto"
                }
              }
            }
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/txs/{id}": {
      "get": {
        "tags": [
          "txs"
        ],
        "summary": "Query a pending Ark tx",
        "description": "Placeholder for a future query endpoint. For now this is a stub that mirrors `ArkService.GetPendingTx` when it becomes available.",
        "operationId": "get_tx",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ark txid",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Pending tx status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PendingTxResponseDto"
                }
              }
            }
          },
          "501": {
            "description": "Not implemented",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/v1/txs/{id}/finalize": {
      "post": {
        "tags": [
          "txs"
        ],
        "summary": "Finalize a pending Ark tx",
        "operationId": "finalize_tx",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Ark txid previously returned by POST /txs",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FinalizeTxRequestDto"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Finalized"
          },
          "400": {
            "description": "Bad request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "502": {
            "description": "Upstream error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConfirmRegistrationRequestDto": {
        "type": "object",
        "required": [
          "intent_id"
        ],
        "properties": {
          "intent_id": {
            "type": "string"
          }
        }
      },
      "EstimateIntentFeeRequestDto": {
        "type": "object",
        "required": [
          "input_vtxo_ids",
          "outputs"
        ],
        "properties": {
          "input_vtxo_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "outputs": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OutputDto"
            }
          }
        }
      },
      "EstimateIntentFeeResponseDto": {
        "type": "object",
        "required": [
          "fee_sats",
          "fee_rate_sats_per_vb"
        ],
        "properties": {
          "fee_rate_sats_per_vb": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "fee_sats": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "FinalizeTxRequestDto": {
        "type": "object",
        "properties": {
          "final_checkpoint_txs": {
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "OutputDto": {
        "type": "object",
        "description": "One output destination: exactly one of `vtxo_script` or `onchain_address`\nmust be set.",
        "required": [
          "amount"
        ],
        "properties": {
          "amount": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "onchain_address": {
            "type": [
              "string",
              "null"
            ]
          },
          "vtxo_script": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "PendingTxResponseDto": {
        "type": "object",
        "required": [
          "ark_txid",
          "status"
        ],
        "properties": {
          "ark_txid": {
            "type": "string"
          },
          "status": {
            "type": "string"
          }
        }
      },
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 problem+json body.",
        "required": [
          "title",
          "status",
          "detail"
        ],
        "properties": {
          "detail": {
            "type": "string",
            "description": "Human-readable explanation specific to this occurrence."
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "description": "HTTP status code for this occurrence of the problem.",
            "minimum": 0
          },
          "title": {
            "type": "string",
            "description": "Short, human-readable summary of the problem type."
          }
        }
      },
      "RegisterIntentRequestDto": {
        "type": "object",
        "required": [
          "proof",
          "message"
        ],
        "properties": {
          "delegate_pubkey": {
            "type": "string",
            "description": "Optional delegate pubkey (hex-encoded compressed). Empty string if absent."
          },
          "message": {
            "type": "string",
            "description": "Canonical intent message (JSON text per the BIP-322 proof)."
          },
          "proof": {
            "type": "string",
            "description": "BIP-322 proof PSBT (hex)."
          }
        }
      },
      "RegisterIntentResponseDto": {
        "type": "object",
        "required": [
          "intent_id"
        ],
        "properties": {
          "intent_id": {
            "type": "string"
          }
        }
      },
...
      "SubmitTxRequestDto": {
        "type": "object",
        "required": [
          "signed_ark_tx"
        ],
        "properties": {
          "checkpoint_txs": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Checkpoint transactions in hex (optional)."
          },
          "signed_ark_tx": {
            "type": "string",
            "description": "Signed Ark virtual tx, hex encoded."
          }
        }
      },
      "SubmitTxResponseDto": {
        "type": "object",
        "required": [
          "ark_txid"
        ],
        "properties": {
          "ark_txid": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "macaroon"
      }
    }
  },
  "security": [
    {
      "bearer": []
    }
  ],
...
}